requirements.txt
# Telegram Bot API
pyTelegramBotAPI==4.14.0

# HTTP и асинхронные операции
aiohttp==3.9.1
aiosqlite==0.19.0

# Парсинг HTML
beautifulsoup4==4.12.3
lxml==5.1.0

# Утилиты
pytz==2024.1
python-dotenv==1.0.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0

